requirements.txt
discord.py[voice]
spotipy
PyNaCl